homework3/requirements.txt
aiohttp==3.7.4.post0
alembic==1.5.7
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
colorama==0.4.4
dataclasses==0.6
greenlet==1.0.0
idna==3.1
loguru==0.5.3
Mako==1.1.4
MarkupSafe==1.1.1
multidict==5.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.0
typing-extensions==3.7.4.3
win32-setctime==1.0.3
yarl==1.6.3
